+++ examples/valid_commit_type_click.py
@@ -7,6 +7,12 @@ import sys
 import click
 from icecream import ic
 
+from incolume.py.githooks.core import (
+    CONTEXT_SETTINGS_CLICK,
+    __package_name__,
+    __version__,
+)
+
 commit_types = [
     'chore',
     'docs',
@@ -18,7 +24,12 @@ commit_types = [
 ]
 
 
-@click.command()
+@click.command(context_settings=CONTEXT_SETTINGS_CLICK)
+@click.version_option(
+    package_name=__package_name__,
+    version=__version__,
+    prog_name='validate_commit_type_click',
+)
 @click.argument(
     'commit_msg_file',
     default='',